Fix release workflow
For now, we don't actually create a Github release and push the assets
since that creates some complexities with the OS build matrix.
So we just upload the assets (they are ZIPed + their original
compression) and the once we want to create a release that step will
have to be done manually.